Move most of the packs directory into the src directory
Resolves #6
