+++ app/javascript/src/admin/components/dashboard/index.vue
