+++ app/javascript/src/admin/components/musicians/_filters.vue
