+++ app/javascript/src/admin/components/musicians/edit.vue
