+++ app/javascript/src/admin/components/musicians/index.vue
