+++ app/javascript/src/admin/components/shared/_pagination.vue
