+++ app/javascript/src/admin/components/shared/_submit_tag.vue
