+++ app/javascript/src/admin/components/shared/layout.vue
