+++ app/javascript/src/admin/event_bus.js
