+++ app/javascript/src/admin/filters/strings.js
