+++ app/javascript/src/admin/helpers/index.js
