+++ app/javascript/src/admin/vuex/index.js
