+++ app/javascript/src/components/errors/500.vue
