+++ app/javascript/src/components/errors/index.vue
