+++ app/javascript/src/components/musicians/show.vue
