+++ app/javascript/src/filters/strings.js
