+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20171218221703) do
+ActiveRecord::Schema.define(version: 2017_12_18_221703) do
 
   create_table "musicians", force: :cascade do |t|
     t.datetime "created_at", null: false
